package creation and some refactorization
