+++ app/src/main/java/com/to_do_list/app/AddToDo.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.widget.*
 import androidx.appcompat.app.AppCompatActivity
 
-class AddToDO : AppCompatActivity() {
+class AddToDo : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
