+++ app/src/main/java/com/to_do_list/app/ListOnClickListener.kt
@@ -1,5 +1,7 @@
 package com.to_do_list.app
 
+import com.to_do_list.app.common.entities.ToDoList
+
 interface ListOnClickListener {
     fun onListClickListener(data: ToDoList)
