+++ app/src/main/java/com/to_do_list/app/TodoAddUpdateDeleteActivity.kt
@@ -5,6 +5,7 @@ import android.view.View
 import android.widget.*
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import com.to_do_list.app.common.entities.ToDo
 
 class TodoAddUpdateDeleteActivity : AppCompatActivity(), View.OnClickListener {
