+++ app/src/main/java/com/to_do_list/app/common/entities/ToDo.kt
@@ -1,29 +1,29 @@
-package com.to_do_list.app
+package com.to_do_list.app.common.entities
 
-import android.graphics.drawable.Drawable
 import android.os.Parcel
 import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "ToDo", foreignKeys = arrayOf(
-    ForeignKey(
-        entity = ToDoList::class,
-        parentColumns = arrayOf("id"),
-        childColumns = arrayOf("list"),
-        onDelete = ForeignKey.CASCADE
+@Entity(
+    tableName = "ToDo", foreignKeys = arrayOf(
+        ForeignKey(
+            entity = ToDoList::class,
+            parentColumns = arrayOf("id"),
+            childColumns = arrayOf("list"),
+            onDelete = ForeignKey.CASCADE
+        )
     )
-))
+)
 data class ToDo(
     @PrimaryKey(autoGenerate = true)
-    val id : Int,
+    val id: Int,
     val description: String,
     val completionState: Int = 0,
     val assignedPeople: String,
-    val list : Int
-)
-    : Parcelable {
+    val list: Int
+) : Parcelable {
     constructor(parcel: Parcel) : this(
         parcel.readInt(),
         parcel.readString()!!,
